Put command-line arguments related functions in a separate file
